+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 `1.0.6`
 
+- Changed `Fields.IgnoreAccess.get` with field name as parameter to look for the field in the class hierarchy.
+- Changed `Fields.IgnoreAccess.set` with field name as parameter to look for the field in the class hierarchy.
+- Changed `Fields.IgnoreAccess.getStatic` with field name as parameter to look for the field in the class hierarchy because they are accessible through derived classes too.
+- Changed `Fields.IgnoreAccess.setStatic` with field name as parameter to look for the field in the class hierarchy because they are accessible through derived classes too.
+
 ---
 
 `1.0.5`